bigquery-agent-service/requirements.txt
fastapi
uvicorn
google-cloud-aiplatform
google-cloud-bigquery
pydantic
python-dateutil
google-auth
